refactor: reorganize structure and set dao interface
