+++ go.mod
@@ -3,10 +3,10 @@ module github.com/gdgtoledo/linneo
 go 1.12
 
 require (
-	github.com/elastic/go-elasticsearch/v8 v8.0.0-20190906153242-7540ece56511 // indirect
+	github.com/elastic/go-elasticsearch/v8 v8.0.0-20190906153242-7540ece56511
 	github.com/gin-gonic/gin v1.4.0
-	github.com/sirupsen/logrus v1.4.2 // indirect
-	go.elastic.co/apm/module/apmelasticsearch v1.5.0 // indirect
+	github.com/sirupsen/logrus v1.4.2
+	go.elastic.co/apm/module/apmelasticsearch v1.5.0
 	go.elastic.co/apm/module/apmgin v1.5.0
 	go.elastic.co/apm/module/apmlogrus v1.5.0 // indirect
 )
