+++ src/elastic/delete.go
@@ -4,8 +4,8 @@ import (
 	"github.com/gdgtoledo/linneo/src/dao"
 )
 
-// Delete by id Dao implementation in elastic
-func (p Params) Delete() (dao.Response, error) {
+// Delete by id to Elastic Dao
+func (eDao DataObjectModel) Delete(id dao.ID) (dao.Response, error) {
 	var r dao.Response
 	var err error
 	return r, err
