--- src/elastic/params.go
@@ -1,7 +1,0 @@
-package elastic
-
-// Params for an elastic action
-type Params struct {
-	Query string
-	ID    string
-}
