+++ src/elastic/search.go
@@ -4,15 +4,14 @@ import (
 	"github.com/gdgtoledo/linneo/src/dao"
 )
 
-// Search Dao implementation to elastic
-func (p Params) Search() (dao.Response, error) {
+// Search to Elastic Dao
+func (eDao DataObjectModel) Search(query dao.Query) (dao.Response, error) {
 	var r dao.Response
 	var err error
-+
+
 	if esClientError != nil {
 		return r, err
 	}
-
 
 	return r, err
 }
