+++ src/main.go
@@ -1,19 +1,19 @@
 package main
 
 import (
+	dao "github.com/gdgtoledo/linneo/src/elastic"
 	"github.com/gdgtoledo/linneo/src/routes"
-	"github.com/gdgtoledo/linneo/src/service"
 	"github.com/gin-gonic/gin"
 	"go.elastic.co/apm/module/apmgin"
 )
 
 func handleSearch(c *gin.Context) {
-	r, _ := service.Search("my query")
+	r, _ := dao.Search("my query")
 }
 
 func handleDelete(c *gin.Context) {
 	id := c.Param("id")
-	r, _ := service.Delete(id)
+	r, _ := dao.Delete(id)
 }
 
 func setupRouter() *gin.Engine {
